+++ src/components/manage/Contents/ContentsUploadModal.jsx
@@ -267,7 +267,14 @@ class ContentsUploadModal extends Component {
                     <Table.Row className="upload-row" key={file.name}>
                       <Table.Cell>{file.name}</Table.Cell>
                       <Table.Cell>
-                        <FormattedRelativeDate date={file.lastModifiedDate} />
+                        {file.lastModifiedDate ? (
+                          <FormattedRelativeDate date={file.lastModifiedDate} />
+                        ) : (
+                          <FormattedMessage
+                            id="Not Found"
+                            defaultMessage="Not Found"
+                          />
+                        )}
                       </Table.Cell>
                       <Table.Cell>
                         {filesize(file.size, { round: 0 })}